* tp/Texinfo/XS/convert/build_html_perl_state.c (build_no_arg_commands_formatting, build_html_files_source_info) (build_html_elements_in_file_count, build_html_translated_names) (build_html_command_formatted_args, build_replaced_substrings), tp/Texinfo/XS/convert/convert_html.c (html_get_tree_root_element) (add_associated_file_info_integer, lookup_associated_file_info) (html_register_opened_section_level, prepare_special_units) (set_special_units_targets_files) (prepare_associated_special_units_targets) (set_root_commands_targets_node_files, html_get_css_elements_classes) (html_attribute_class, prepare_index_entries_targets) (prepare_footnotes_targets, set_heading_commands_targets) (html_prepare_output_units_global_targets, find_file_source_info) (html_reset_files_source_info, html_set_pages_files) (prepare_special_units_directions) (html_prepare_units_directions_files, html_default_format_contents) (html_default_format_end_file, get_links) (html_default_format_navigation_panel) (html_default_format_element_header, html_default_format_element_footer) (convert_heading_command, convert_inline_command, convert_sp_command) (convert_quotation_command, convert_tab_command, convert_unit_type) (default_format_special_body_about, html_converter_init_special_unit) (determine_non_default_special_unit_directions) (html_converter_customize, reset_html_targets_list) (html_initialize_output_state, html_conversion_finalization) (clear_type_explanations, html_reset_converter, html_free_converter) (html_translate_names, destroy_args_formatted) (html_convert_type_update_context, debug_print_html_contexts) (convert_to_html_internal, convert_output_unit) (convert_convert_output_unit_internal, html_convert_convert) (convert_output_output_unit_internal, html_convert_output) (html_node_redirections): use size_t instead of int where the variable could reach a value outside of the int range and also when the index is better represented by a size_t.  Cast to/from int and size_t when needed.

+++ tp/Texinfo/XS/convert/build_html_perl_state.c
@@ -94,7 +94,7 @@ SV *
 build_no_arg_commands_formatting (const CONVERTER *converter)
 {
   HV *no_arg_commands_formatting_hv;
-  int i;
+  size_t i;
 
   dTHX;
 
@@ -303,7 +303,7 @@ html_pass_converter_setup_state (const CONVERTER *converter,
 SV *
 build_html_files_source_info (const FILE_SOURCE_INFO_LIST *files_source_info)
 {
-  int i;
+  size_t i;
   HV *hv;
 
   dTHX;
@@ -468,7 +468,7 @@ HV *
 build_html_elements_in_file_count (
                  FILE_NAME_PATH_COUNTER_LIST *output_unit_files)
 {
-  int i;
+  size_t i;
   HV *hv;
 
   dTHX;
@@ -608,7 +608,7 @@ build_html_translated_names (HV *hv, CONVERTER *converter)
   /* pass all the information for each context for translated commands */
   if (converter->no_arg_formatted_cmd_translated.number)
     {
-      int j;
+      size_t j;
       HV *no_arg_commands_formatting_hv;
       FETCH(no_arg_commands_formatting);
       no_arg_commands_formatting_hv
@@ -728,7 +728,7 @@ SV *
 build_html_command_formatted_args (const HTML_ARGS_FORMATTED *args_formatted)
 {
   AV *av;
-  int i;
+  size_t i;
 
   dTHX;
 
@@ -771,7 +771,7 @@ SV *
 build_replaced_substrings (NAMED_STRING_ELEMENT_LIST *replaced_substrings)
 {
   HV *hv;
-  int i;
+  size_t i;
 
   dTHX;
